Rename dependencies.rs to framework.rs
